LayoutTests/fast/dom/treeOrder-expected.txt

Test the tree ordering.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.treeOrder(document, document, 'ComposedTree') is "equivalent"
PASS internals.treeOrder(body, body, 'ComposedTree') is "equivalent"
PASS internals.treeOrder(document, body, 'ComposedTree') is "less"
PASS internals.treeOrder(body, document, 'ComposedTree') is "greater"
PASS internals.treeOrder(a, a, 'ComposedTree') is "equivalent"
PASS internals.treeOrder(body, a, 'ComposedTree') is "unordered"
PASS internals.treeOrder(a, body, 'ComposedTree') is "unordered"
PASS internals.treeOrder(body, a, 'ComposedTree') is "less"
PASS internals.treeOrder(a, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(a, a, 'ComposedTree') is "equivalent"
PASS internals.treeOrder(body, b, 'ComposedTree') is "less"
PASS internals.treeOrder(b, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(a, b, 'ComposedTree') is "less"
PASS internals.treeOrder(b, a, 'ComposedTree') is "greater"
PASS internals.treeOrder(body, c, 'ComposedTree') is "less"
PASS internals.treeOrder(c, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(a, c, 'ComposedTree') is "less"
PASS internals.treeOrder(c, a, 'ComposedTree') is "greater"
PASS internals.treeOrder(b, c, 'ComposedTree') is "less"
PASS internals.treeOrder(c, b, 'ComposedTree') is "greater"
PASS internals.treeOrder(body, d, 'ComposedTree') is "less"
PASS internals.treeOrder(d, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(body, d, 'ComposedTree') is "less"
PASS internals.treeOrder(d, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(a, d, 'ComposedTree') is "less"
PASS internals.treeOrder(d, a, 'ComposedTree') is "greater"
PASS internals.treeOrder(d, b, 'ComposedTree') is "less"
PASS internals.treeOrder(b, d, 'ComposedTree') is "greater"
PASS internals.treeOrder(d, c, 'ComposedTree') is "less"
PASS internals.treeOrder(c, d, 'ComposedTree') is "greater"
PASS internals.treeOrder(a, d, 'ComposedTree') is "less"
PASS internals.treeOrder(d, a, 'ComposedTree') is "greater"
PASS internals.treeOrder(d, b, 'ComposedTree') is "less"
PASS internals.treeOrder(b, d, 'ComposedTree') is "greater"
PASS internals.treeOrder(d, c, 'ComposedTree') is "less"
PASS internals.treeOrder(c, d, 'ComposedTree') is "greater"
PASS internals.treeOrder(body, f, 'ComposedTree') is "unordered"
PASS internals.treeOrder(f, body, 'ComposedTree') is "unordered"
PASS internals.treeOrder(f, body, 'ComposedTree') is "unordered"
PASS internals.treeOrder(body, f, 'ComposedTree') is "unordered"
PASS internals.treeOrder(body, g, 'ComposedTree') is "less"
PASS internals.treeOrder(g, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(body, h, 'ComposedTree') is "less"
PASS internals.treeOrder(h, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(body, i, 'ComposedTree') is "less"
PASS internals.treeOrder(i, body, 'ComposedTree') is "greater"
PASS internals.treeOrder(i, d, 'ComposedTree') is "less"
PASS internals.treeOrder(d, i, 'ComposedTree') is "greater"
PASS internals.treeOrder(i, b, 'ComposedTree') is "less"
PASS internals.treeOrder(b, i, 'ComposedTree') is "greater"
PASS successfullyParsed is true

TEST COMPLETE
